Move containers into own component
