+++ src/Containers/index.js
@@ -1,0 +1,3 @@
+export { ExpensesContainer } from "./ExpensesContainer";
+export { FriendsContainer } from "./FriendsContainer";
+export { SettingsContainer } from "./SettingsContainer";
